Add a test_data_dir py.test option
We are moving data out of our git repos for multiple reasons.  The
data in the streamcorpus repo was hundreds of megabytes for instance.
But our tests both in development and in our installed environment
require this data.

The default location of this data will be /opt/diffeo/data/<packagename>